third-party/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiobotocore[awscli,boto3]==1.1.2
    # via
    #   -r requirements.in
    #   s3fs
aiohttp==3.7.3
    # via aiobotocore
aioitertools==0.7.1
    # via aiobotocore
alabaster==0.7.12
    # via sphinx
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   markdown-it-py
awscli==1.18.121
    # via aiobotocore
babel==2.9.0
    # via sphinx
backcall==0.2.0
    # via ipython
beautifulsoup4==4.9.3
    # via bs4
bleach==3.3.0
    # via nbconvert
boto3==1.14.44
    # via
    #   -r requirements.in
    #   aiobotocore
botocore==1.17.44
    # via
    #   aiobotocore
    #   awscli
    #   boto3
    #   s3transfer
breathe==4.19.2
    # via
    #   -r requirements.in
    #   exhale
bs4==0.0.1
    # via exhale
certifi==2020.12.5
    # via
    #   matplotlib
    #   requests
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
click==7.1.2
    # via keyname
colorama==0.4.3
    # via awscli
coverxygen==1.5.0
    # via -r requirements.in
cycler==0.10.0
    # via matplotlib
decorator==4.4.2
    # via
    #   ipython
    #   networkx
    #   traitlets
defusedxml==0.7.1
    # via nbconvert
docutils==0.15.2
    # via
    #   awscli
    #   botocore
    #   breathe
    #   myst-parser
    #   sphinx
entrypoints==0.3
    # via nbconvert
exdown==0.7.0
    # via -r requirements.in
exhale==0.2.3
    # via -r requirements.in
frozendict==1.2
    # via -r requirements.in
fsspec==0.8.5
    # via
    #   -r requirements.in
    #   s3fs
gitdb==4.0.5
    # via gitpython
gitpython==3.1.14
    # via -r requirements.in
idna==2.8
    # via
    #   -r requirements.in
    #   requests
    #   yarl
image-slicer==2.1.1
    # via -r requirements.in
imagesize==1.2.0
    # via sphinx
ipykernel==5.5.3
    # via -r requirements.in
ipython-genutils==0.2.0
    # via
    #   nbformat
    #   traitlets
ipython==7.16.1
    # via
    #   -r requirements.in
    #   ipykernel
iterdub==0.2.0
    # via -r requirements.in
iterpop==0.3.4
    # via -r requirements.in
j2cli[yaml]==0.3.10
    # via -r requirements.in
jedi==0.18.0
    # via ipython
jinja2==2.11.3
    # via
    #   j2cli
    #   myst-parser
    #   nbconvert
    #   sphinx
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via nbformat
jupyter-core==4.7.1
    # via
    #   -r requirements.in
    #   jupyter-client
    #   nbconvert
    #   nbformat
jupyter_client==6.1.12
    # via
    #   -r requirements.in
    #   ipykernel
keyname==0.4.1
    # via -r requirements.in
kiwisolver==1.3.1
    # via matplotlib
lxml==4.6.3
    # via exhale
markdown-it-py==0.6.2
    # via
    #   -r requirements.in
    #   mdit-py-plugins
    #   myst-parser
markdown2==2.3.9
    # via -r requirements.in
markupsafe==1.1.1
    # via jinja2
matplotlib==3.3.2
    # via
    #   -r requirements.in
    #   seaborn
mdit-py-plugins==0.2.6
    # via
    #   markdown-it-py
    #   myst-parser
mistune==0.8.4
    # via nbconvert
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
myst-parser==0.13.5
    # via -r requirements.in
nbconvert==5.6.1
    # via -r requirements.in
nbformat==5.1.3
    # via nbconvert
networkx==2.5
    # via -r requirements.in
numpy==1.19.5
    # via
    #   -r requirements.in
    #   iterdub
    #   matplotlib
    #   opencv-python
    #   pandas
    #   patsy
    #   scipy
    #   seaborn
    #   statsmodels
oauthlib==3.1.0
    # via requests-oauthlib
opencv-python==4.5.1.48
    # via -r requirements.in
osfclient==0.0.5
    # via -r requirements.in
packaging==20.9
    # via
    #   bleach
    #   sphinx
pandas==1.1.2
    # via
    #   -r requirements.in
    #   seaborn
    #   statsmodels
pandocfilters==1.4.3
    # via nbconvert
parso==0.8.2
    # via jedi
patsy==0.5.1
    # via statsmodels
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==7.2.0
    # via
    #   image-slicer
    #   matplotlib
prompt-toolkit==3.0.18
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pyasn1==0.4.8
    # via rsa
pygments==2.8.1
    # via
    #   ipython
    #   nbconvert
    #   sphinx
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.17.3
    # via jsonschema
pysocks==1.7.1
    # via requests
python-dateutil==2.8.1
    # via
    #   botocore
    #   jupyter-client
    #   matplotlib
    #   pandas
python-slugify==4.0.1
    # via -r requirements.in
pytz==2021.1
    # via
    #   babel
    #   pandas
pyyaml==5.3.1
    # via
    #   awscli
    #   j2cli
    #   myst-parser
pyzmq==22.0.3
    # via jupyter-client
requests-oauthlib==1.3.0
    # via tweepy
requests[socks]==2.25.1
    # via
    #   osfclient
    #   requests-oauthlib
    #   sphinx
    #   tweepy
rsa==4.5
    # via awscli
s3fs==0.5.1
    # via -r requirements.in
s3transfer==0.3.4
    # via
    #   awscli
    #   boto3
scipy==1.5.4
    # via
    #   -r requirements.in
    #   seaborn
    #   statsmodels
seaborn==0.11.0
    # via -r requirements.in
six==1.15.0
    # via
    #   bleach
    #   breathe
    #   cycler
    #   exhale
    #   jsonschema
    #   osfclient
    #   patsy
    #   python-dateutil
    #   traitlets
    #   tweepy
smmap==3.0.5
    # via gitdb
snowballstemmer==2.1.0
    # via sphinx
soupsieve==2.2.1
    # via beautifulsoup4
sphinx-rtd-theme==0.5.0
    # via -r requirements.in
sphinx==3.1.2
    # via
    #   -r requirements.in
    #   breathe
    #   exhale
    #   myst-parser
    #   sphinx-rtd-theme
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==1.0.3
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.4
    # via sphinx
statsmodels==0.12.2
    # via -r requirements.in
testpath==0.4.4
    # via nbconvert
text-unidecode==1.3
    # via python-slugify
toml==0.10.1
    # via -r requirements.in
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.50.2
    # via
    #   -r requirements.in
    #   osfclient
traitlets==4.3.3
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   nbconvert
    #   nbformat
tweepy==3.10.0
    # via -r requirements.in
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   aioitertools
urllib3==1.25.11
    # via
    #   botocore
    #   requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
wrapt==1.12.1
    # via aiobotocore
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
